feat: trip controller
